Refactor 2.0.0
Refactored the entire project. Using the MinecraftSessionService instead of LoginListener to allow for more compatibility accross versions. Added a game profile wrapper as a parameter to the AsyncPlayerProfileCreationEvent so that users can view and manipulate the players current properties inside that event. MinecraftSessionService does not provide an InetAddress on versions prior to 1.11, thus the address on the AsyncPlayerProfileCreationEvent is now an Optional.